Implement Issue #3: Extract data fetching logic to useAnalysisData hook
